refactor(snmptraps): migrate config component to v2 structure
- Move interface to config/def/ alongside TrapsConfig types
- Move implementation to config/impl/ with Requires/Provides/NewComponent
- Add config/fx/ with fxutil.ProvideComponentConstructor
- Update import paths in forwarder, listener, and server
