+++ comp/snmptraps/config/def/component.go
