+++ comp/snmptraps/config/def/config.go
