+++ comp/snmptraps/config/impl/service.go
@@ -13,9 +13,33 @@ import (
 
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/comp/core/hostname"
-	trapsconf "github.com/DataDog/datadog-agent/comp/snmptraps/config"
+	trapsconf "github.com/DataDog/datadog-agent/comp/snmptraps/config/def"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
+
+// Requires defines the dependencies for the config component.
+type Requires struct {
+	fx.In
+
+	Conf      config.Component
+	HNService hostname.Component
+}
+
+// Provides defines the output of the config component.
+type Provides struct {
+	fx.Out
+
+	Comp trapsconf.Component
+}
+
+// NewComponent creates a new config component.
+func NewComponent(reqs Requires) (Provides, error) {
+	comp, err := newService(reqs.Conf, reqs.HNService)
+	if err != nil {
+		return Provides{}, err
+	}
+	return Provides{Comp: comp}, nil
+}
 
 type configService struct {
 	conf *trapsconf.TrapsConfig
